Lambdas/Lenet5Lambda/image_processor/requirements.txt
scipy==1.0.0
Pillow==5.0.0
boto3==1.5.18